apply min/max via static style
